Fix Possible run-time error due to casting int to short
Values that were declared as type short were changed to be declared as type int.
Short casted values returned from calling, for example,
DataInputStream.readUnsignedByte() were removed and the type of the variable used
to store these returns was changed to type int to hold this value, as
DataInputStream.readUnsignedByte() returns an int (see DataInputStream).

All data types in the Class File structure and its members are meant to be
interpreted as unsigned. Per the documentation, the DataInputStream class is
suggested to be used to read u1, u2, or u4 values. Methods used, such as
readUnsignedShort() return an integer that represents the unsigned value of
the 2 bytes read (see DataInputStream). Since all scalar values are treated as signed
in the jvm, narrowing this value back down to a short changes the sign bit. This could
lead to a negative value, which is incorrect.

The above becomes a serious issue when the short casted value was meant to represent an
index into the constant pool. A negative value is not a valid index (see Java Class File).
Therefore, all values should not be casted back down to a short.

References:
Java Class File, https://docs.oracle.com/javase/specs/jvms/se19/html/jvms-4.html#jvms-4.2.1
DataInputStream, Java API, https://docs.oracle.com/en/java/javase/19/docs/api/java.base/java/io/DataInputStream.html

+++ src/main/java/j_parser/ClassFile.java
@@ -12,41 +12,41 @@ import j_parser.interfaces.Constant;
 
 public class ClassFile{
     private final int magicNumber;
-    private final short minorVersion;
-    private final short majorVersion;
+    private final int minorVersion;
+    private final int majorVersion;
     private final ConstantPool cPool;
-    private final short constantCount;
+    private final int constantCount;
 /*
-    private final short accessFlags;
-    private final short thisClass;
+    private final int accessFlags;
+    private final int thisClass;
 
-    private final short superClass;
-    private final short interfacesCount;
-    private short[] interfaces;
-    private short fieldsCount;
-    private HashMap<short, Field> fields;
-    private short methodsCount;
-    private HashMap<short, Method> methods;
-    private short attributesCount;
-    private HashMap<short, Attribute> attributes;
+    private final int superClass;
+    private final int interfacesCount;
+    private int[] interfaces;
+    private int fieldsCount;
+    private HashMap<int, Field> fields;
+    private int methodsCount;
+    private HashMap<int, Method> methods;
+    private int attributesCount;
+    private HashMap<int, Attribute> attributes;
 */
     public ClassFile(File classFile) throws IOException{
         DataInputStream fileStream = new DataInputStream(new FileInputStream(classFile));
 
         this.magicNumber = fileStream.readInt();
-        this.minorVersion = (short)fileStream.readUnsignedShort();
-        this.majorVersion = (short)fileStream.readUnsignedShort();
+        this.minorVersion = fileStream.readUnsignedShort();
+        this.majorVersion = fileStream.readUnsignedShort();
         this.cPool = new ConstantPool(fileStream); 
         this.constantCount = this.cPool.getConstantCount();
    }
 
     public int getMagicNumber(){return this.magicNumber;}
     
-    public short getMinorVersion(){return this.minorVersion;}
+    public int getMinorVersion(){return this.minorVersion;}
     
-    public short getMajorVersion(){return this.majorVersion;}
+    public int getMajorVersion(){return this.majorVersion;}
     
-    public short getConstantCount(){return this.constantCount;}
+    public int getConstantCount(){return this.constantCount;}
     
     public ArrayList<String> getStrings(){
         return this.cPool.getStrings();
