+++ src/main/java/j_parser/types/constants/ConstantUTF.java
@@ -5,16 +5,16 @@ import j_parser.utils.ConstantTag;
 
 public final class ConstantUTF implements Constant{
     private final int TAG = ConstantTag.UTF.TAG;
-    private final short LEN;
+    private final int LEN;
     private final byte[] BYTES;
     private final String COMMON_NAME = ConstantTag.UTF.name();
 
-    public ConstantUTF(short len, byte[] bytes){
+    public ConstantUTF(int len, byte[] bytes){
         this.LEN = len;
         this.BYTES = bytes;
     }
 
-    public short getLen(){
+    public int getLen(){
         return this.LEN;
     }
